修正代码，具体见partner\hot-deploy\partner\document\测试\系统测试情况-刘继煌.xlsx

+++ hot-deploy/partner/src/org/ofbiz/partner/scm/pricemgr/Utils.java
@@ -221,7 +221,7 @@ public class Utils {
 			tempEntry.set("volume", entryValue.getBigDecimal("volume"));
 			tempEntry.set("price", entryValue.getBigDecimal("price"));
 			tempEntry.set("entrysum", entryValue.getBigDecimal("entrysum"));
-			tempEntry.set("sort", entryValue.getBigDecimal("sort"));
+			tempEntry.set("sort", entryValue.getInteger("sort"));
 			tempEntry.create();
 		}
 	}
